Separate Out Control Plane Create
Makes life a lot easer when form values aren't interspersed with view
logic.
